remove repeat defined VariadicValue type
